Revert "Merge remote-tracking branch 'remote/GitHubApps' into GitHubApps"
This reverts commit c53cc110b8d807f62fdfeaa7df19e1532d050007, reversing
changes made to 0c9e413d420a4725738644ea5b13af6ec102d456.

+++ Octokit/Helpers/ApiUrls.cs
@@ -328,7 +328,7 @@ namespace Octokit
         /// Returns the <see cref="Uri"/> that returns a single installation of the authenticated application.
         /// </summary>
         /// <returns></returns>
-        public static Uri Installation(long installationId)
+        public static Uri Installation(int installationId)
         {
             return "app/installations/{0}".FormatUri(installationId);
         }
